+++ starsim/distributions.py
@@ -8,7 +8,7 @@ import scipy.stats as sps
 import starsim as ss
 import matplotlib.pyplot as pl
 
-__all__ = ['find_dists', 'link_dists', 'make_dist', 'dist_list', 'Dists', 'Dist']
+__all__ = ['link_dists', 'make_dist', 'dist_list', 'Dists', 'Dist']
 
 
 def str2int(string, modulo=1_000_000):
@@ -21,25 +21,13 @@ def str2int(string, modulo=1_000_000):
     return int.from_bytes(string.encode(), byteorder='big') % modulo
 
 
-def find_dists(obj, verbose=False, **kwargs):
-    """ Find all Dist objects in a parent object """
-    out = sc.objdict()
-    tree = sc.iterobj(obj, depthfirst=False, flatten=True, **kwargs)
-    if verbose: print(f'Found {len(tree)} objects')
-    for trace,val in tree.items():
-        if isinstance(val, Dist):
-            out[trace] = val
-            if verbose: print(f'  {trace} is a dist ({len(out)})')
-    return out
-
-
 def link_dists(obj, sim, module=None, overwrite=False, init=False, **kwargs):
     """ Link distributions to the sim and the module; used in module.initialize() and people.initialize() """
     if module is None and isinstance(obj, ss.Module):
         module = obj
-    dists = ss.find_dists(obj, **kwargs) # Important that this comes first, before the sim is linked to the dist!
+    dists = ss.find_objs(Dist, obj, **kwargs) # Important that this comes first, before the sim is linked to the dist!
     for key,val in dists.items():
-        if isinstance(val, ss.Dist):
+        if isinstance(val, ss.Dist): # TODO: should be able to skip this if statement since should always be true?
             val.link_sim(sim, overwrite=overwrite)
             val.link_module(module, overwrite=overwrite)
             if init: # Usually this is false since usually these are initialized centrally by the sim
@@ -97,7 +85,7 @@ class Dists(sc.prettyobj):
         skip = id(sim.people._states) if sim is not None else None
         
         # Find and initialize the distributions
-        self.dists = find_dists(obj, skip=skip)
+        self.dists = ss.find_objs(Dist, obj, skip=skip)
         for trace,dist in self.dists.items():
             if not dist.initialized or force:
                 dist.initialize(trace=trace, seed=base_seed, sim=sim, force=force)